requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appnope
asttokens
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
comm
cssutils==2.11.1
colormath
datasets==2.20.0
debugpy
decorator
dill==0.3.8
distro==1.9.0
exceptiongroup
executing
filelock==3.15.4
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
fsspec==2024.5.0
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.24.0
idna==3.7
importlib_metadata
ipykernel
ipython
ipywidgets==8.1.3
itsdangerous==2.2.0
jedi
Jinja2==3.1.4
joblib==1.4.2
jupyter_client
jupyter_core
jupyterlab_widgets==3.0.11
lxml==5.2.2
MarkupSafe==3.0.2
matplotlib-inline
more-itertools==10.3.0
multidict==6.0.5
multiprocess==0.70.16
nest_asyncio
numpy==1.26.4
openai==1.13.3
outcome==1.3.0.post0
packaging
pandas==2.2.2
parso
pexpect
pickleshare
pillow==10.4.0
platformdirs
prompt_toolkit
proto-plus==1.24.0
protobuf==4.25.3
psutil
ptyprocess
pure-eval
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil
pytz==2024.1
PyYAML==6.0.1
pyzmq
regex==2024.5.15
requests==2.32.3
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.22.0
six
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.36
stack-data
threadpoolctl==3.5.0
tokenizers==0.19.1
tornado
tqdm==4.66.4
traitlets
transformers==4.42.3
trio==0.26.0
trio-websocket==0.11.1
typing_extensions
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
wcwidth
websocket-client==1.8.0
Werkzeug==3.0.6
widgetsnbextension==4.0.11
setuptools==59.8.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
zipp
lap
scikit_image==0.24.0
open_clip_torch
anthropic
lpips

